# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Stackverse Bookmarks API
  version: 1.0.0
  description: |
    The contract every Stackverse backend implements and every frontend consumes.

    Authentication is a bearer JWT supplied by the gateway (token relay) — clients
    behind the gateway never handle tokens themselves. Business rules that don't fit
    an OpenAPI document live in `docs/SPEC.md`.

    Versioning is URI-based (`/api/v{n}/...`) and evolves per operation: breaking
    changes introduce a successor version that coexists with the deprecated
    predecessor, which keeps answering with `Deprecation`, `Sunset`, and
    `Link rel="successor-version"` headers. The v1 → v2 bookmarks listing is the
    permanent exhibit of this policy (see `docs/SPEC.md`).
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8000
    description: through the gateway (session cookie, token relayed for you)
  - url: http://localhost:8080
    description: backend directly (you must supply the bearer token yourself)
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/bookmarks/{id}/reports:
    post:
      operationId: reportBookmark
      tags: [moderation]
      summary: Report a public bookmark
      description: |
        Any authenticated user may report a bookmark that is public and visible to
        them. At most one `open` report per user per bookmark (`409`); a new report
        is allowed once the previous one is resolved or withdrawn. Private or
        hidden bookmarks yield `404` — existence is not disclosed.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportInput'
      responses:
        '201':
          description: Report created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The caller already has an open report on this bookmark.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/reports:
    get:
      operationId: listMyReports
      tags: [moderation]
      summary: The caller's own reports
      description: |
        Every report the caller has filed, newest first — the reporter's
        feedback loop. `status` shows what moderation decided; the resolution
        fields (`resolvedBy`, `resolvedAt`, `resolutionNote`) carry the
        disposition once a moderator acts.
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ReportStatus'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: A page of the caller's reports, newest first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/reports/{id}:
    put:
      operationId: updateMyReport
      tags: [moderation]
      summary: Update the caller's own open report
      description: |
        The reporter may revise `reason` and `comment` while the report is
        `open`. Another user's report yields `404` — existence is not
        disclosed. A resolved report yields `409`.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportInput'
      responses:
        '200':
          description: The updated report.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The report is not open.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    delete:
      operationId: withdrawReport
      tags: [moderation]
      summary: Withdraw the caller's own open report
      description: |
        Removes the report entirely; a withdrawn report no longer blocks filing
        a new one on the same bookmark. Another user's report yields `404` —
        existence is not disclosed. A resolved report yields `409`.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Report withdrawn.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The report is not open.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /api/v1/admin/reports:
    get:
      operationId: listReports
      tags: [moderation]
      summary: Moderation queue (moderator)
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ReportStatus'
            default: open
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: A page of reports, oldest open first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/reports/{id}:
    put:
      operationId: resolveReport
      tags: [moderation]
      summary: Resolve or revise a report (moderator)
      description: |
        `dismissed` leaves the bookmark alone; `actioned` hides it and auto-resolves
        every other open report on the same bookmark as `actioned` with the same
        resolver and note. Decisions are revisable: any target status is accepted —
        `dismissed` ↔ `actioned` changes the disposition (`actioned` applies its
        side effects as usual), and `open` re-opens the report, clearing the
        resolution fields (any `note` sent with `open` is ignored). Moving away
        from `actioned` never restores the bookmark; use the bookmark status
        endpoint for that.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportResolutionInput'
      responses:
        '200':
          description: The resolved report.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/admin/bookmarks/{id}/status:
    put:
      operationId: setBookmarkStatus
      tags: [moderation]
      summary: Hide or restore a bookmark (moderator)
      description: 'Restoring sets `status: active` and leaves `visibility` untouched.'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookmarkStatusInput'
      responses:
        '200':
          description: The bookmark with its new status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bookmark'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages (public, paginated)
      security: []
      parameters:
        - name: key
          in: query
          description: Exact message key.
          schema:
            $ref: '#/components/schemas/MessageKey'
        - name: q
          in: query
          description: Case-insensitive substring match over key and text.
          schema:
            type: string
            maxLength: 200
        - name: language
          in: query
          schema:
            $ref: '#/components/schemas/LanguageCode'
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: A page of messages.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessagePage'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      operationId: createMessage
      tags: [messages]
      summary: Create a message (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageInput'
      responses:
        '201':
          description: Created. `Location` header points at the new resource.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/messages/bundle:
    get:
      operationId: getMessageBundle
      tags: [messages]
      summary: Flat key→text bundle for one resolved language
      description: |
        Returns every message key with its text in the resolved language
        (`lang` parameter → `Accept-Language` → `en`); keys missing in that
        language fall back to their `en` text. `Content-Language` reports the
        language actually served. This is what frontends load on startup and
        on language switch.
      security: []
      parameters:
        - name: lang
          in: query
          description: Explicit language override; unsupported values fall back silently.
          schema:
            $ref: '#/components/schemas/LanguageCode'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: The bundle.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Content-Language:
              description: Language actually served.
              schema:
                $ref: '#/components/schemas/LanguageCode'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageBundle'
        '304':
          $ref: '#/components/responses/NotModified'
  /api/v1/messages/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getMessage
      tags: [messages]
      summary: Get a message (public)
      security: []
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: The message.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '304':
          $ref: '#/components/responses/NotModified'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: updateMessage
      tags: [messages]
      summary: Replace a message (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageInput'
      responses:
        '200':
          description: The updated message.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message (admin only)
      responses:
        '204':
          description: Deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access token issued by Keycloak, relayed by the gateway. Backends validate
        issuer, signature (via JWKS), expiry, and audience; identity comes from the
        `preferred_username` claim, roles from `realm_access.roles`. Operations
        marked "admin only" require the `admin` role.
  parameters:
    IfNoneMatch:
      name: If-None-Match
      in: header
      description: Entity tag from a previous response; an unchanged representation returns `304`.
      schema:
        type: string
  headers:
    ETag:
      description: Entity tag of the current representation.
      schema:
        type: string
    CacheControl:
      description: 'Always `no-cache`: responses may be stored but must be revalidated before reuse.'
      schema:
        type: string
...
  schemas:
    Visibility:
      type: string
      enum: [private, public]
    Tag:
      type: string
      pattern: '^[a-z0-9-]{1,30}$'
    LanguageCode:
      type: string
      pattern: '^[a-z]{2}$'
      description: Two-letter lowercase ISO 639-1 code.
    MessageKey:
      type: string
      pattern: '^[a-z0-9-]+(\.[a-z0-9-]+)*$'
      maxLength: 150
      description: Dot-separated lowercase segments, e.g. `ui.nav.my-bookmarks`.
    MessageInput:
      type: object
      required: [key, language, text]
      properties:
        key:
          $ref: '#/components/schemas/MessageKey'
        language:
          $ref: '#/components/schemas/LanguageCode'
        text:
          type: string
          minLength: 1
          maxLength: 2000
        description:
          type: string
          maxLength: 1000
          description: Context for translators.
    Message:
      allOf:
        - $ref: '#/components/schemas/MessageInput'
        - type: object
          required: [id, createdAt, updatedAt]
          properties:
            id:
              type: string
              format: uuid
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    MessagePage:
      type: object
      required: [items, page, size, totalItems, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Message'
        page:
          type: integer
        size:
          type: integer
        totalItems:
          type: integer
          format: int64
        totalPages:
          type: integer
    MessageBundle:
      type: object
      required: [language, messages]
      properties:
        language:
          $ref: '#/components/schemas/LanguageCode'
        messages:
          type: object
          additionalProperties:
            type: string
          description: Flat key → localized text map.
    BookmarkInput:
      type: object
      required: [url, title]
      properties:
        url:
          type: string
          format: uri
          maxLength: 2000
          description: Absolute http(s) URL.
        title:
          type: string
          minLength: 1
          maxLength: 200
        notes:
          type: string
          maxLength: 4000
        tags:
          type: array
          maxItems: 10
          items:
            $ref: '#/components/schemas/Tag'
          description: Normalized server-side — trimmed, lowercased, deduplicated.
        visibility:
          $ref: '#/components/schemas/Visibility'
          default: private
    BookmarkStatus:
      type: string
      enum: [active, hidden]
      description: Server-managed; changes only through moderation.
    Bookmark:
      allOf:
        - $ref: '#/components/schemas/BookmarkInput'
        - type: object
          required: [id, owner, status, createdAt, updatedAt, visibility, tags]
          properties:
            id:
              type: string
              format: uuid
            owner:
              type: string
              description: Username of the creator. Immutable.
            status:
              $ref: '#/components/schemas/BookmarkStatus'
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
...
    BookmarkStatusInput:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/BookmarkStatus'
        note:
          type: string
          maxLength: 1000
    ReportReason:
      type: string
      enum: [spam, offensive, broken-link, other]
    ReportStatus:
      type: string
      enum: [open, dismissed, actioned]
    ReportInput:
      type: object
      required: [reason]
      properties:
        reason:
          $ref: '#/components/schemas/ReportReason'
        comment:
          type: string
          maxLength: 1000
    Report:
      allOf:
        - $ref: '#/components/schemas/ReportInput'
        - type: object
          required: [id, bookmarkId, reporter, status, createdAt]
          properties:
            id:
              type: string
              format: uuid
            bookmarkId:
              type: string
              format: uuid
            reporter:
              type: string
            status:
              $ref: '#/components/schemas/ReportStatus'
            createdAt:
              type: string
              format: date-time
            resolvedBy:
              type: string
            resolvedAt:
              type: string
              format: date-time
            resolutionNote:
              type: string
    ReportPage:
      type: object
      required: [items, page, size, totalItems, totalPages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Report'
        page:
          type: integer
        size:
          type: integer
        totalItems:
          type: integer
          format: int64
        totalPages:
          type: integer
    ReportResolutionInput:
      type: object
      required: [resolution]
      properties:
        resolution:
          type: string
          enum: [open, dismissed, actioned]
        note:
          type: string
          maxLength: 1000
...
    Problem:
      type: object
      description: RFC 9457 problem details. `errors` carries field-level validation failures.
      properties:
        type:
          type: string
          format: uri
          default: 'about:blank'
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        errors:
          type: array
          items:
            type: object
            required: [field, messageKey, message]
            properties:
              field:
                type: string
              messageKey:
                $ref: '#/components/schemas/MessageKey'
              message:
                type: string
                description: Localized per the language resolution rules (`lang` → `Accept-Language` → `en`).
  responses:
    BadRequest:
      description: Validation failure.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: Missing or invalid token (or, through the gateway, no session).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Authenticated, but missing the required role.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource does not exist — or is not visible to the caller.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotModified:
      description: The representation matches `If-None-Match`; the body is empty.
    Conflict:
      description: A message with this `(key, language)` pair already exists.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'